backend/requirements.txt
flask
flask-cors
flask-session
flask-sqlalchemy
pymysql
numpy<1.19.0,>=1.16.0
scipy
python-Levenshtein
thefuzz
tensorflow-gpu
sent2vec
pyautogui
paho-mqtt
